Fourth version giving very similar results much faster.
Reverse looping order, fix for sloppy test, now much faster.
The innermost loop with dynamic looping bounds now deals with
the lowest-order term in each pass, greatly reducing looping.
This makes the program finish in less than 15 seconds instead
of taking a bit over 2 and a half minutes on my best laptop.

Fix the case where the bounds test ended when the next number
(i + 1) has the same error level as the current. Expand up to
avoid missing optimal numbers. This restores result quality,
after using the bounds testing and sizing for the lowest term
exposed the problem. A remaining difference is that the order
of tests now differs and that a different variation with the
same key numbers now ends up becoming the result of running.

(Note from the 10th: Similar but different results come from
a few items being missed, because bounds for looping limited
too much on the lower end. The result is worse but the error
measures equally high with single-precision floating point.)