Sixth version restoring the result in my 2021-08-04 article.
(The article which is linked to in README.md.)

Speed up the use of the new algorithm by half,
to under half the old performance using broken
fast code (the versions with similar results).

It's slower but it doesn't miss items, and now
the result in my 08-04 article is back. Though
items could still be missed in other potential
cases. This is a placeholder, for better code.