rework bounds-narrowing code before abandoning ratio positions
More complex bounds-narrowing code which leaves too-large
intervals for linear search for the program to be usable.

Narrows down to a power of two,
first, then switches to finding
the average of upper-middle, or
middle-lower, bounds repeatedly
until done.

Simplified in next commit, which replaces integer (n-1)/n
positions with simple floating-point 0.0 to 1.0 positions
and seems to make it work, by narrowing down to one final
number.