Seventh very experimental version, better result for double.
Weight tests, optimize a little, tweak result table format.

Remove probe_at(), use run_one().

Try again with the recursive version of
the new algorithm but it still fails in
the same way, despite some tweaks which
by accident improve the result.

This version contains good ideas and runs the fastest of any
so far which gives a good result, but I doubt it is correct.